new glx_init.c
