+++ src/public/glx.c
@@ -6,75 +6,11 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/09 11:46:33 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/03/20 11:29:39 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/03/21 01:15:18 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "glx.h"
-
-static t_glx_prv	*_glx_init_private(size_t update_lim)
-{
-	t_glx_prv	*prv;
-
-	prv = ft_calloc(1, sizeof(t_glx_prv));
-	if (!prv)
-		return (NULL);
-	prv->update_count = 0;
-	prv->update_lim = update_lim;
-	prv->error = _glx_error;
-	return (prv);
-}
-
-static bool	_glx_init_is_successed(t_glx *glx)
-{
-	if (!glx || !glx->user || !glx->_)
-	{
-		ft_printf("glx_init: memory_allocation failed.\n");
-		ft_printf("glx:%p\tglx->user:%p\tglx->_:%p\n", glx, glx->user, glx->_);
-		free(glx);
-		free(glx->user);
-		free(glx->_);
-		return (false);
-	}
-	return (true);
-}
-
-/**
- * @brief glxを初期化します。
- *
- * @param title
- * @param win_w
- * @param win_h
- * @param update_lim
- * @return t_glx*
- */
-t_glx	*glx_init(char *title, int win_w, int win_h, size_t update_lim)
-{
-	t_glx	*glx;
-
-	glx = ft_calloc(1, sizeof(t_glx));
-	glx->user = ft_calloc(1, sizeof(t_glx_user));
-	glx->_ = _glx_init_private(update_lim);
-	if (!_glx_init_is_successed(glx))
-		exit(EXIT_FAILURE);
-	glx->frame_count = 0;
-	glx->mlx = mlx_init();
-	glx->win = mlx_new_window(glx->mlx, win_w, win_h, title);
-	glx->imgc = 0;
-	glx->run = glx_run;
-	glx->hook = glx_hook;
-	glx->load_img = glx_load_img;
-	glx->put_img = glx_put_img;
-	glx->text = glx_text;
-	glx->quit = glx_quit;
-	glx->btn = glx_btn;
-	glx->btnp = glx_btnp;
-	glx->cls = glx_cls;
-	_glx_key_state_init(glx);
-	_glx_key_just_state_init(glx);
-	set_glx(glx);
-	return (glx);
-}
 
 /**
  * @brief glxを終了します。内部で確保したメモリを開放します。
